Generate scenes

+++ src/bsdfs/DielectricBSDF.cpp
@@ -23,7 +23,7 @@ void DielectricBSDF::generateSample(const Computation& _comp, ngl::Vec4& sampleD
     ngl::Vec3 direction;
 
     // Check conditions for reflection vs refraction, and compute the corresponding direction.
-    if (cannot_refract || reflectance(cos_theta, refraction_ratio) > randomDouble(0,1))
+    if (cannot_refract || reflectance(cos_theta, refraction_ratio) > randomFloat(0,1))
         direction = unit_direction.reflect(_comp.normal.toVec3());
     else
         direction = refract(unit_direction, _comp.normal.toVec3(), refraction_ratio);
@@ -48,7 +48,7 @@ void DielectricBSDF::evaluateSample(const Computation& _comp, const ngl::Vec4& s
     bool cannot_refract = refraction_ratio * sin_theta > 1.0;
     ngl::Vec3 expectedDirection;
 
-    if (cannot_refract || reflectance(cos_theta, refraction_ratio) > randomDouble(0,1))
+    if (cannot_refract || reflectance(cos_theta, refraction_ratio) > randomFloat(0,1))
         expectedDirection = unit_direction.reflect(_comp.normal.toVec3());
     else
         expectedDirection = refract(unit_direction, _comp.normal.toVec3(), refraction_ratio);
